# backend/app.py
from flasgger import Schema, Swagger, SwaggerView, fields, swag_from
from flask import Flask, jsonify, request, render_template, session, send_from_directory, Blueprint
from flask_cors import CORS
from flask_graphql import GraphQLView
from flask_jwt_extended import (
    JWTManager,
    create_access_token,
    get_jwt_identity,
    get_raw_jwt,
    jwt_required,
)
from flask_restful import Api, Resource
from sqlalchemy import text

from backend.config import Config
from backend.helpers import QUERY_KEYS, SALARY_KEYS, STATES
from backend.models import Salary, db
from backend.schema import schema
from flask_sqlalchemy import SQLAlchemy
from flasgger import Swagger, swag_from, SwaggerView, Schema, fields
from flask_swagger_ui import get_swaggerui_blueprint


app = Flask(__name__)
CORS(app)
app.config.from_object(Config)

jwt = JWTManager(app)
db.app = app
db.init_app(app)
swagger = Swagger(app)
api = Api(app)

@app.route('/static/<path:path>/')
def send_static(path):
    return send_from_directory('static', path)

SWAGGER_URL = '/swagger'
API_URL = '/static/swagger.yml'
swaggerui_blueprint = get_swaggerui_blueprint(
    SWAGGER_URL,
    API_URL,
    config={
        'app_name': "Salary Database"
    }
)
app.register_blueprint(swaggerui_blueprint, url_prefix=SWAGGER_URL)
request_api = Blueprint('request_api', __name__)
app.register_blueprint(request_api)


@app.route("/", methods=["GET"])
def home():
    """
    Home page route
    """
    form_values = {}
    for query in QUERY_KEYS:
        value = request.args.get(query, session.get(query, "")).strip()
        form_values[query] = value
        session[query] = value
    ppl = [
        {key: salary.__dict__[key] for key in SALARY_KEYS}
        for salary in Salary.query.limit(10).all()
    ]
    return render_template("index.html", states=STATES, form_values=form_values)


@app.route("/data", methods=["GET"])
def employer():
    params = []
    for d in range(0, 4):
        index = str(d + 1)
        field = request.args.get("field" + index, None)
        value = request.args.get("value" + index, None)
        if field and value:
            params.append(field.upper())
            params.append(value.upper())
        elif not (field or value):
            params.append("''")
            params.append("")
        else:
            return jsonify({"message": "A parameter is missing."}), 400

    data = db.engine.execute(
        """SELECT * FROM salary WHERE (%s = '%s' AND %s = '%s' AND %s = '%s' AND %s = '%s')
        """
        % (
            params[0],
            params[1],
            params[2],
            params[3],
            params[4],
            params[5],
            params[6],
            params[7],
        )
    )
    final_result = [dict(i) for i in data]
    return jsonify({"results": final_result}), 200

    db.session.commit()
    return jsonify(message="success"), 200


# GraphQL page route
app.add_url_rule(
    "/graphql", view_func=GraphQLView.as_view("graphql", schema=schema, graphiql=True),
)


@app.route("/table", methods=["GET"])
def table():
    """
    Endpoint used by table in frontend to fetch salary data 

    Retrieves filter queries from Session and returns sorted and 
    paginated data according to the limit, offset, sort, and order
    query params

    Query string:
        - limit(str) - max number of results to return
        - offset(str) - starting point of data
        - sort(str) - column to sort data by. default: prevailing_wage
        - order(str) - how to order column being sorted(asc or desc). default: asc
    returns:
        matched salary data in form of dict with keys:
            - total - total number of matched results. not the same as limit
            - rows - matched salary data in current page as a list
    """
    max_results = int(request.args.get("limit"))
    offset = int(request.args.get("offset"))
    page = offset // max_results
    sort_by = request.args.get("sort", "prevailing_wage")
    order = request.args.get("order", "asc")
    fuzzy_query = {}
    for query in QUERY_KEYS:
        fuzzy_query[query] = "%{}%".format(session[query])
    matched = Salary.query.filter(
        db.or_(Salary.employer_name.like(fuzzy_query["employer"])),
        db.or_(Salary.job_title.like(fuzzy_query["title"])),
        db.or_(Salary.employer_city.like(fuzzy_query["city"])),
        db.or_(Salary.employer_state.like(fuzzy_query["state"])),
        db.or_(
            db.cast(db.extract("year", Salary.employment_start_date), db.String).like(
                fuzzy_query["year"]
            )
        ),
    )

    rows = [
        {key: salary.__dict__[key] for key in SALARY_KEYS}
        for salary in matched.order_by(getattr(getattr(Salary, sort_by), order)())
        .paginate(page + 1, max_results, False)
        .items
    ]
    return jsonify({"total": matched.count(), "rows": rows})


@app.route("/about", methods=["GET"])
def about():
    """
    About page route
    """
    return render_template("about.html")


@app.route("/api", methods=["GET"])
def api_ref():
    """
    API Reference page route
    """
    return render_template("api.html")


if __name__ == "__main__":
    app.run(debug=True, host="0.0.0.0")